+++ modules/randomwalk/ext/RandomWalksLib/SparseSolverEigenBiCGSTAB.cpp
@@ -27,7 +27,7 @@ std::vector<double> SparseSolverEigenBiCGSTAB::solve_Ax_b( SparseMatrix<double> 
 		xmat[uidx[i]] = val;
 	}
 
-	for (int i=0; i<seeds->size(); i++)
+	for (size_t i=0; i<seeds->size(); i++)
 	{
 		if((*labels)[i] == active_label)
 			xmat[(*seeds)[i]] = 1.0;
